la api crea un usuario

+++ models/claim.go
@@ -5,6 +5,6 @@ import "github.com/golang-jwt/jwt"
 type Claim struct {
 	Email       string `json:"email"`
 	TipoUsuario string `json:"tipoUsuario"`
-	IdUser      uint   `json:"iduser"`
+	IdUser      uint64 `json:"iduser"`
 	jwt.StandardClaims
 }
