+++ models/educacion.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type Educacion struct {
-	ID                    uint      `json:"id" gorm:"primary_key"`
-	UsuarioID             uint      `json:"usuarioID"` // foreignKey
+	ID                    uint64    `json:"id" gorm:"primary_key"`
+	UsuarioID             uint64    `json:"usuarioID"` // foreignKey
 	InstituacionEducativa string    `json:"instituacionEducativa"`
 	Titulo                string    `json:"titulo"`
 	DiciplinaAcademica    string    `json:"diciplinaAcademica"`
